feat: Rename server run command to just run
Closes #976
