+++ Dockerfile
@@ -37,4 +37,4 @@ EXPOSE 8080 8082
 
 ENTRYPOINT ["/usr/bin/influxdb_iox"]
 
-CMD ["server", "run"]
+CMD ["run"]
