+++ docker/Dockerfile.iox
@@ -21,4 +21,4 @@ EXPOSE 8080 8082
 
 ENTRYPOINT ["influxdb_iox"]
 
-CMD ["server", "run"]
+CMD ["run"]
