+++ docs/env.example
@@ -7,7 +7,7 @@
 # The full list of available configuration values can be found by in
 # the command line help (e.g. `env: INFLUXDB_IOX_DB_DIR=`):
 #
-# ./influxdb_iox server --help
+# ./influxdb_iox run --help
 #
 #
 # The identifier for the server. Used for writing to object storage and as
